Extract state to roomContext provider, give context to room component
