+++ frontend/src/App.js
@@ -1,12 +1,15 @@
 import './App.css';
 import Room from './components/Room'
+import RoomProvider from './providers/RoomProvider';
 
 const App = function() {
 
   return (
     <div className="Parrot-perch">
       <h1>Hello World!</h1>
-      <Room />
+      <RoomProvider>
+        <Room />
+      </RoomProvider>
     </div>
 
   );
