+++ frontend/src/components/ChatInput.scss
@@ -1,0 +1,5 @@
+#chat-submit {
+  border-radius: 0.5rem;
+  width: 20em;
+  height: 2.3em;
+}
