+++ frontend/src/components/Room.jsx
@@ -1,38 +1,16 @@
-import io from 'socket.io-client';
-import { useState, useEffect } from 'react';
+import { useContext } from 'react';
 import Join from './JoinRoom';
 import View from './ViewRoom';
-import Chat from './components/Chat';
+import Chat from './Chat';
+import { roomContext } from '../providers/RoomProvider';
 
 function Room() {
-  const [socket, setSocket] = useState();
-  const [isViewing, setIsViewing] = useState(false);
-  const [room, setRoom] = useState({
-    name: '',
-    channel: '',
-    users: []
-  });
+  const { room, setRoom, isViewing, socket } = useContext(roomContext);
 
   const handleJoin = function (e) {
     e.preventDefault();
     socket.emit('createOrJoinRoom', { room });
-  }
-
-  useEffect(() => {
-    const socket = io();
-    setSocket(socket);
-    socket.on('connect', () => {
-      console.log('connected');
-    });
-    socket.on('serveRoom', (res) => {
-      const room = res.room;
-      setRoom((oldRoom) => {
-        return { ...oldRoom, name: room.name, channel: room.channel, users: room.users };
-      });
-      setIsViewing(true);
-    });
-    return () => socket.disconnect();
-  }, [])
+  };
 
   return (
     <>
